refactor: API base URL 분리
