+++ src/api/config.ts
@@ -1,0 +1,2 @@
+export const baseUrl =
+  'http://techcourse-lv2-alb-974870821.ap-northeast-2.elb.amazonaws.com';
