+++ src/ui/components/ProductSection/ProductSection.tsx
@@ -1,12 +1,11 @@
+import React from 'react';
 import Title from '../Title/Title';
 import Dropdown from '../Dropdown/Dropdown';
 import ProductList from '../ProductList/ProductList';
 import { DropdownContainer, Section } from './ProductSection.styles';
 import { CATEGORY, SORT_PRICE } from '../../../constants/productConfig';
 import { SortType, CategoryType } from '../../../types/product';
-import React from 'react';
-import { ProductElement } from '../../../App';
-
+import { ProductElement } from '../../../types/product';
 interface ProductSectionProps {
   data: any;
   sort: SortType;
@@ -28,7 +27,7 @@ function ProductSection({
 }: ProductSectionProps) {
   return (
     <Section>
-      <Title title="bpple 상품 목록" />
+      <Title title="우테코 상품 목록" />
       <DropdownContainer>
         <Dropdown value={category} options={CATEGORY} onChange={onFilter} />
         <Dropdown value={sort} options={SORT_PRICE} onChange={onSort} />
